refactor: 로그인 관련 객체들 네이밍, 파일 위치 수정
