+++ Kustaurant/Kustaurant/Data/Dto/AccessTokenResponse.swift
@@ -1,0 +1,12 @@
+//
+//  AccessTokenResponse.swift
+//  Kustaurant
+//
+//  Created by peppermint100 on 8/21/24.
+//
+
+import Foundation
+
+struct AccessTokenResponse: Codable {
+    let accessToken: String
+}
