+++ Kustaurant/Kustaurant/Domain/Entities/KuUser.swift
@@ -11,17 +11,9 @@ enum SocialLoginProvider: String, Codable {
     case naver, apple
 }
 
-struct SocialLoginUser: Codable {
+struct KuUser: Codable {
     let id: String
     let accessToken: String
     let provider: SocialLoginProvider
 }
 
-struct NaverMeResponse: Codable {
-    let resultcode: String
-    let response: Response
-    
-    struct Response: Codable {
-        let id: String
-    }
-}
